back/pyproject.toml
[project]
name = 'back'
version = '0.1.1'
description = 'EMIT Backend API'
readme = 'README.md'
requires-python = '>=3.12'
dependencies = [
    'fastapi>=0.115.0',
    'uvicorn[standard]>=0.32.0',
    'tortoise-orm>=0.21.0',
    'pydantic>=2.9.0',
    "passlib>=1.7.4",
    "python-multipart>=0.0.20",
    "bcrypt<5.0",
    "python-jose[cryptography]>=3.5.0",
    "starlette>=0.48.0",
]
